Extract shared test helpers into test_helpers_test.go
